kill , fix tests
