+++ src/core/builtin.c
@@ -39,275 +39,6 @@ void minim_load_builtin(MinimEnv *env, const char *name, MinimObjectType type, .
     env_intern_sym(env, name, obj);
 }
 
-void minim_load_builtins(MinimEnv *env)
-{
-    // Variable / Control
-    minim_load_builtin(env, "def-values", MINIM_OBJ_SYNTAX, minim_builtin_def_values);
-    minim_load_builtin(env, "set!", MINIM_OBJ_SYNTAX, minim_builtin_setb);
-    minim_load_builtin(env, "if", MINIM_OBJ_SYNTAX, minim_builtin_if);
-    minim_load_builtin(env, "let-values", MINIM_OBJ_SYNTAX, minim_builtin_let_values);
-    minim_load_builtin(env, "let*-values", MINIM_OBJ_SYNTAX, minim_builtin_letstar_values);
-    minim_load_builtin(env, "begin", MINIM_OBJ_SYNTAX, minim_builtin_begin);
-    minim_load_builtin(env, "quote", MINIM_OBJ_SYNTAX, minim_builtin_quote);
-    minim_load_builtin(env, "quasiquote", MINIM_OBJ_SYNTAX, minim_builtin_quasiquote);
-    minim_load_builtin(env, "unquote", MINIM_OBJ_SYNTAX, minim_builtin_unquote);
-    minim_load_builtin(env, "lambda", MINIM_OBJ_SYNTAX, minim_builtin_lambda);
-    minim_load_builtin(env, "exit", MINIM_OBJ_FUNC, minim_builtin_exit);
-    minim_load_builtin(env, "call/cc", MINIM_OBJ_SYNTAX, minim_builtin_callcc);
-
-    // Modules
-    minim_load_builtin(env, "%export", MINIM_OBJ_SYNTAX, minim_builtin_export);
-    minim_load_builtin(env, "%import", MINIM_OBJ_SYNTAX, minim_builtin_import);
-
-    // Transforms
-    minim_load_builtin(env, "def-syntaxes", MINIM_OBJ_SYNTAX, minim_builtin_def_syntaxes);
-    minim_load_builtin(env, "quote-syntax", MINIM_OBJ_SYNTAX, minim_builtin_syntax);
-    minim_load_builtin(env, "syntax?", MINIM_OBJ_FUNC, minim_builtin_syntaxp);
-    minim_load_builtin(env, "unwrap", MINIM_OBJ_FUNC, minim_builtin_unwrap);
-    minim_load_builtin(env, "syntax-case", MINIM_OBJ_SYNTAX, minim_builtin_syntax_case);
-    minim_load_builtin(env, "datum->syntax", MINIM_OBJ_FUNC, minim_builtin_to_syntax);
-    minim_load_builtin(env, "syntax", MINIM_OBJ_SYNTAX, minim_builtin_template);
-
-    // Errors
-    minim_load_builtin(env, "error", MINIM_OBJ_FUNC, minim_builtin_error);
-    minim_load_builtin(env, "argument-error", MINIM_OBJ_FUNC, minim_builtin_argument_error);
-    minim_load_builtin(env, "arity-error", MINIM_OBJ_FUNC, minim_builtin_arity_error);
-    minim_load_builtin(env, "syntax-error", MINIM_OBJ_FUNC, minim_builtin_syntax_error);
-    
-    // Miscellaneous
-    minim_load_builtin(env, "values", MINIM_OBJ_FUNC, minim_builtin_values);
-    minim_load_builtin(env, "equal?", MINIM_OBJ_FUNC, minim_builtin_equalp);
-    minim_load_builtin(env, "eqv?", MINIM_OBJ_FUNC, minim_builtin_eqvp);
-    minim_load_builtin(env, "eq?", MINIM_OBJ_FUNC, minim_builtin_eqp);
-    minim_load_builtin(env, "symbol?", MINIM_OBJ_FUNC, minim_builtin_symbolp);
-    minim_load_builtin(env, "printf", MINIM_OBJ_FUNC, minim_builtin_printf);
-    minim_load_builtin(env, "void", MINIM_OBJ_FUNC, minim_builtin_void);
-    minim_load_builtin(env, "version", MINIM_OBJ_FUNC, minim_builtin_version);
-    minim_load_builtin(env, "symbol-count", MINIM_OBJ_FUNC, minim_builtin_symbol_count);
-    minim_load_builtin(env, "dump-symbols", MINIM_OBJ_FUNC, minim_builtin_dump_symbols);
-    minim_load_builtin(env, "def-print-method", MINIM_OBJ_FUNC, minim_builtin_def_print_method);
-
-    // Arity
-    minim_load_builtin(env, "procedure?", MINIM_OBJ_FUNC, minim_builtin_procedurep);
-    minim_load_builtin(env, "procedure-arity", MINIM_OBJ_FUNC, minim_builtin_procedure_arity);
-
-    // Boolean
-    env_intern_sym(env, "true", minim_true);
-    env_intern_sym(env, "false", minim_false);
-    minim_load_builtin(env, "bool?", MINIM_OBJ_FUNC, minim_builtin_boolp);
-    minim_load_builtin(env, "not", MINIM_OBJ_FUNC, minim_builtin_not);
-
-    // Numbers
-    minim_load_builtin(env, "number?", MINIM_OBJ_FUNC, minim_builtin_numberp);
-    minim_load_builtin(env, "zero?", MINIM_OBJ_FUNC, minim_builtin_zerop);
-    minim_load_builtin(env, "positive?", MINIM_OBJ_FUNC, minim_builtin_positivep);
-    minim_load_builtin(env, "negative?", MINIM_OBJ_FUNC, minim_builtin_negativep);
-    minim_load_builtin(env, "even?", MINIM_OBJ_FUNC, minim_builtin_evenp);
-    minim_load_builtin(env, "odd?", MINIM_OBJ_FUNC, minim_builtin_oddp);
-    minim_load_builtin(env, "exact?", MINIM_OBJ_FUNC, minim_builtin_exactp);
-    minim_load_builtin(env, "inexact?", MINIM_OBJ_FUNC, minim_builtin_inexactp);
-    minim_load_builtin(env, "integer?", MINIM_OBJ_FUNC, minim_builtin_integerp);
-    minim_load_builtin(env, "nan?", MINIM_OBJ_FUNC, minim_builtin_nanp);
-    minim_load_builtin(env, "infinite?", MINIM_OBJ_FUNC, minim_builtin_infinitep);
-    
-    minim_load_builtin(env, "=", MINIM_OBJ_FUNC, minim_builtin_eq);
-    minim_load_builtin(env, ">", MINIM_OBJ_FUNC, minim_builtin_gt);
-    minim_load_builtin(env, "<", MINIM_OBJ_FUNC, minim_builtin_lt);
-    minim_load_builtin(env, ">=", MINIM_OBJ_FUNC, minim_builtin_gte);
-    minim_load_builtin(env, "<=", MINIM_OBJ_FUNC, minim_builtin_lte);
-
-    minim_load_builtin(env, "exact", MINIM_OBJ_FUNC, minim_builtin_to_exact);
-    minim_load_builtin(env, "inexact", MINIM_OBJ_FUNC, minim_builtin_to_inexact);
-    
-    minim_load_builtin(env, "inf", MINIM_OBJ_INEXACT, INFINITY);
-    minim_load_builtin(env, "-inf", MINIM_OBJ_INEXACT, -INFINITY);
-    minim_load_builtin(env, "nan", MINIM_OBJ_INEXACT, NAN);
-
-    // Character
-    minim_load_builtin(env, "char?", MINIM_OBJ_FUNC, minim_builtin_charp);
-    minim_load_builtin(env, "char=?", MINIM_OBJ_FUNC, minim_builtin_char_eqp);
-    minim_load_builtin(env, "char>?", MINIM_OBJ_FUNC, minim_builtin_char_gtp);
-    minim_load_builtin(env, "char<?", MINIM_OBJ_FUNC, minim_builtin_char_ltp);
-    minim_load_builtin(env, "char>=?", MINIM_OBJ_FUNC, minim_builtin_char_gtep);
-    minim_load_builtin(env, "char<=?", MINIM_OBJ_FUNC, minim_builtin_char_ltep);
-    minim_load_builtin(env, "char-ci=?", MINIM_OBJ_FUNC, minim_builtin_char_ci_eqp);
-    minim_load_builtin(env, "char-ci>?", MINIM_OBJ_FUNC, minim_builtin_char_ci_gtp);
-    minim_load_builtin(env, "char-ci<?", MINIM_OBJ_FUNC, minim_builtin_char_ci_ltp);
-    minim_load_builtin(env, "char-ci>=?", MINIM_OBJ_FUNC, minim_builtin_char_ci_gtep);
-    minim_load_builtin(env, "char-ci<=?", MINIM_OBJ_FUNC, minim_builtin_char_ci_ltep);
-    minim_load_builtin(env, "char-alphabetic?", MINIM_OBJ_FUNC, minim_builtin_char_alphabeticp);
-    minim_load_builtin(env, "char-numeric?", MINIM_OBJ_FUNC, minim_builtin_char_numericp);
-    minim_load_builtin(env, "char-whitespace?", MINIM_OBJ_FUNC, minim_builtin_char_whitespacep);
-    minim_load_builtin(env, "char-upper-case?", MINIM_OBJ_FUNC, minim_builtin_char_upper_casep);
-    minim_load_builtin(env, "char-lower-case?", MINIM_OBJ_FUNC, minim_builtin_char_lower_casep);
-    minim_load_builtin(env, "integer->char", MINIM_OBJ_FUNC, minim_builtin_int_to_char);
-    minim_load_builtin(env, "char->integer", MINIM_OBJ_FUNC, minim_builtin_char_to_int);
-    minim_load_builtin(env, "char-upcase", MINIM_OBJ_FUNC, minim_builtin_char_upcase);
-    minim_load_builtin(env, "char-downcase", MINIM_OBJ_FUNC, minim_builtin_char_downcase);
-
-    // String
-    minim_load_builtin(env, "string?", MINIM_OBJ_FUNC, minim_builtin_stringp);
-    minim_load_builtin(env, "make-string", MINIM_OBJ_FUNC, minim_builtin_make_string);
-    minim_load_builtin(env, "string", MINIM_OBJ_FUNC, minim_builtin_string);
-    minim_load_builtin(env, "string-length", MINIM_OBJ_FUNC, minim_builtin_string_length);
-    minim_load_builtin(env, "string-ref", MINIM_OBJ_FUNC, minim_builtin_string_ref);
-    minim_load_builtin(env, "string-set!", MINIM_OBJ_FUNC, minim_builtin_string_setb);
-    minim_load_builtin(env, "string-copy", MINIM_OBJ_FUNC, minim_builtin_string_copy);
-    minim_load_builtin(env, "string-fill!", MINIM_OBJ_FUNC, minim_builtin_string_fillb);
-    minim_load_builtin(env, "substring", MINIM_OBJ_FUNC, minim_builtin_substring);
-    minim_load_builtin(env, "string->symbol", MINIM_OBJ_FUNC, minim_builtin_string_to_symbol);
-    minim_load_builtin(env, "symbol->string", MINIM_OBJ_FUNC, minim_builtin_symbol_to_string);
-    minim_load_builtin(env, "string->number", MINIM_OBJ_FUNC, minim_builtin_string_to_number);
-    minim_load_builtin(env, "number->string", MINIM_OBJ_FUNC, minim_builtin_number_to_string);
-    minim_load_builtin(env, "format", MINIM_OBJ_FUNC, minim_builtin_format);
-    
-    // Pair
-    minim_load_builtin(env, "cons", MINIM_OBJ_FUNC, minim_builtin_cons);
-    minim_load_builtin(env, "pair?", MINIM_OBJ_FUNC, minim_builtin_consp);
-    minim_load_builtin(env, "car", MINIM_OBJ_FUNC, minim_builtin_car);
-    minim_load_builtin(env, "cdr", MINIM_OBJ_FUNC, minim_builtin_cdr);
-    env_intern_sym(env, "null", minim_null);
-
-    minim_load_builtin(env, "caar", MINIM_OBJ_FUNC, minim_builtin_caar);
-    minim_load_builtin(env, "cadr", MINIM_OBJ_FUNC, minim_builtin_cadr);
-    minim_load_builtin(env, "cdar", MINIM_OBJ_FUNC, minim_builtin_cdar);
-    minim_load_builtin(env, "cddr", MINIM_OBJ_FUNC, minim_builtin_cddr);
-
-    minim_load_builtin(env, "caaar", MINIM_OBJ_FUNC, minim_builtin_caaar);
-    minim_load_builtin(env, "caadr", MINIM_OBJ_FUNC, minim_builtin_caadr);
-    minim_load_builtin(env, "cadar", MINIM_OBJ_FUNC, minim_builtin_cadar);
-    minim_load_builtin(env, "caddr", MINIM_OBJ_FUNC, minim_builtin_caddr);
-    minim_load_builtin(env, "cdaar", MINIM_OBJ_FUNC, minim_builtin_cdaar);
-    minim_load_builtin(env, "cdadr", MINIM_OBJ_FUNC, minim_builtin_cdadr);
-    minim_load_builtin(env, "cddar", MINIM_OBJ_FUNC, minim_builtin_cddar);
-    minim_load_builtin(env, "cdddr", MINIM_OBJ_FUNC, minim_builtin_cdddr);
-
-    minim_load_builtin(env, "caaaar", MINIM_OBJ_FUNC, minim_builtin_caaaar);
-    minim_load_builtin(env, "caaadr", MINIM_OBJ_FUNC, minim_builtin_caaadr);
-    minim_load_builtin(env, "caadar", MINIM_OBJ_FUNC, minim_builtin_caadar);
-    minim_load_builtin(env, "caaddr", MINIM_OBJ_FUNC, minim_builtin_caaddr);
-    minim_load_builtin(env, "cadaar", MINIM_OBJ_FUNC, minim_builtin_cadaar);
-    minim_load_builtin(env, "cadadr", MINIM_OBJ_FUNC, minim_builtin_cadadr);
-    minim_load_builtin(env, "caddar", MINIM_OBJ_FUNC, minim_builtin_caddar);
-    minim_load_builtin(env, "cadddr", MINIM_OBJ_FUNC, minim_builtin_cadddr);
-    minim_load_builtin(env, "cdaaar", MINIM_OBJ_FUNC, minim_builtin_cdaaar);
-    minim_load_builtin(env, "cdaadr", MINIM_OBJ_FUNC, minim_builtin_cdaadr);
-    minim_load_builtin(env, "cdadar", MINIM_OBJ_FUNC, minim_builtin_cdadar);
-    minim_load_builtin(env, "cdaddr", MINIM_OBJ_FUNC, minim_builtin_cdaddr);
-    minim_load_builtin(env, "cddaar", MINIM_OBJ_FUNC, minim_builtin_cddaar);
-    minim_load_builtin(env, "cddadr", MINIM_OBJ_FUNC, minim_builtin_cddadr);
-    minim_load_builtin(env, "cdddar", MINIM_OBJ_FUNC, minim_builtin_cdddar);
-    minim_load_builtin(env, "cddddr", MINIM_OBJ_FUNC, minim_builtin_cddddr);
-
-    // List
-    minim_load_builtin(env, "list", MINIM_OBJ_FUNC, minim_builtin_list);
-    minim_load_builtin(env, "list?", MINIM_OBJ_FUNC, minim_builtin_listp);
-    minim_load_builtin(env, "null?", MINIM_OBJ_FUNC, minim_builtin_nullp);
-    minim_load_builtin(env, "length", MINIM_OBJ_FUNC, minim_builtin_length);
-    minim_load_builtin(env, "append", MINIM_OBJ_FUNC, minim_builtin_append);
-    minim_load_builtin(env, "reverse", MINIM_OBJ_FUNC, minim_builtin_reverse);
-    minim_load_builtin(env, "list-ref", MINIM_OBJ_FUNC, minim_builtin_list_ref);
-    minim_load_builtin(env, "map", MINIM_OBJ_FUNC, minim_builtin_map);
-    minim_load_builtin(env, "andmap", MINIM_OBJ_FUNC, minim_builtin_andmap);
-    minim_load_builtin(env, "ormap", MINIM_OBJ_FUNC, minim_builtin_ormap);
-    minim_load_builtin(env, "apply", MINIM_OBJ_FUNC, minim_builtin_apply);
-
-    // Hash table
-    minim_load_builtin(env, "hash", MINIM_OBJ_FUNC, minim_builtin_hash);
-    minim_load_builtin(env, "hash?", MINIM_OBJ_FUNC, minim_builtin_hashp);
-    minim_load_builtin(env, "hash-key?", MINIM_OBJ_FUNC, minim_builtin_hash_keyp);
-    minim_load_builtin(env, "hash-ref", MINIM_OBJ_FUNC, minim_builtin_hash_ref);
-    minim_load_builtin(env, "hash-remove", MINIM_OBJ_FUNC, minim_builtin_hash_remove);
-    minim_load_builtin(env, "hash-set", MINIM_OBJ_FUNC, minim_builtin_hash_set);
-    minim_load_builtin(env, "hash-set!", MINIM_OBJ_FUNC, minim_builtin_hash_setb);
-    minim_load_builtin(env, "hash-remove!", MINIM_OBJ_FUNC, minim_builtin_hash_removeb);
-    minim_load_builtin(env, "hash->list", MINIM_OBJ_FUNC, minim_builtin_hash_to_list);
-
-    // Vector
-    minim_load_builtin(env, "vector", MINIM_OBJ_FUNC, minim_builtin_vector);
-    minim_load_builtin(env, "make-vector", MINIM_OBJ_FUNC, minim_builtin_make_vector);
-    minim_load_builtin(env, "vector?", MINIM_OBJ_FUNC, minim_builtin_vectorp);
-    minim_load_builtin(env, "vector-length", MINIM_OBJ_FUNC, minim_builtin_vector_length);
-    minim_load_builtin(env, "vector-ref", MINIM_OBJ_FUNC, minim_builtin_vector_ref);
-    minim_load_builtin(env, "vector-set!", MINIM_OBJ_FUNC, minim_builtin_vector_setb);
-    minim_load_builtin(env, "vector->list", MINIM_OBJ_FUNC, minim_builtin_vector_to_list);
-    minim_load_builtin(env, "list->vector", MINIM_OBJ_FUNC, minim_builtin_list_to_vector);
-    minim_load_builtin(env, "vector-fill!", MINIM_OBJ_FUNC, minim_builtin_vector_fillb);
-
-    // Sequence
-    minim_load_builtin(env, "sequence", MINIM_OBJ_FUNC, minim_builtin_sequence);
-    minim_load_builtin(env, "sequence?", MINIM_OBJ_FUNC, minim_builtin_sequencep);
-    minim_load_builtin(env, "sequence-first", MINIM_OBJ_FUNC, minim_builtin_sequence_first);
-    minim_load_builtin(env, "sequence-rest", MINIM_OBJ_FUNC, minim_builtin_sequence_rest);
-    minim_load_builtin(env, "sequence-empty?", MINIM_OBJ_FUNC, minim_builtin_sequence_donep);
-
-    // Math
-    minim_load_builtin(env, "+", MINIM_OBJ_FUNC, minim_builtin_add);
-    minim_load_builtin(env, "-", MINIM_OBJ_FUNC, minim_builtin_sub);
-    minim_load_builtin(env, "*", MINIM_OBJ_FUNC, minim_builtin_mul);
-    minim_load_builtin(env, "/", MINIM_OBJ_FUNC, minim_builtin_div);
-    minim_load_builtin(env, "abs", MINIM_OBJ_FUNC, minim_builtin_abs);
-    minim_load_builtin(env, "max", MINIM_OBJ_FUNC, minim_builtin_max);
-    minim_load_builtin(env, "min", MINIM_OBJ_FUNC, minim_builtin_min);
-    minim_load_builtin(env, "modulo", MINIM_OBJ_FUNC, minim_builtin_modulo);
-    minim_load_builtin(env, "remainder", MINIM_OBJ_FUNC, minim_builtin_remainder);
-    minim_load_builtin(env, "quotient", MINIM_OBJ_FUNC, minim_builtin_quotient);
-    minim_load_builtin(env, "numerator", MINIM_OBJ_FUNC, minim_builtin_numerator);
-    minim_load_builtin(env, "denominator", MINIM_OBJ_FUNC, minim_builtin_denominator);
-    minim_load_builtin(env, "gcd", MINIM_OBJ_FUNC, minim_builtin_gcd);
-    minim_load_builtin(env, "lcm", MINIM_OBJ_FUNC, minim_builtin_lcm);
-
-    minim_load_builtin(env, "floor", MINIM_OBJ_FUNC, minim_builtin_floor);
-    minim_load_builtin(env, "ceil", MINIM_OBJ_FUNC, minim_builtin_ceil);
-    minim_load_builtin(env, "trunc", MINIM_OBJ_FUNC, minim_builtin_trunc);
-    minim_load_builtin(env, "round", MINIM_OBJ_FUNC, minim_builtin_round);
-
-    minim_load_builtin(env, "sqrt", MINIM_OBJ_FUNC, minim_builtin_sqrt);
-    minim_load_builtin(env, "exp", MINIM_OBJ_FUNC, minim_builtin_exp);
-    minim_load_builtin(env, "log", MINIM_OBJ_FUNC, minim_builtin_log);
-    minim_load_builtin(env, "pow", MINIM_OBJ_FUNC, minim_builtin_pow);
-
-    minim_load_builtin(env, "sin", MINIM_OBJ_FUNC, minim_builtin_sin);
-    minim_load_builtin(env, "cos", MINIM_OBJ_FUNC, minim_builtin_cos);
-    minim_load_builtin(env, "tan", MINIM_OBJ_FUNC, minim_builtin_tan);
-    minim_load_builtin(env, "asin", MINIM_OBJ_FUNC, minim_builtin_asin);
-    minim_load_builtin(env, "acos", MINIM_OBJ_FUNC, minim_builtin_acos);
-    minim_load_builtin(env, "atan", MINIM_OBJ_FUNC, minim_builtin_atan);
-
-    // Promises
-    minim_load_builtin(env, "delay", MINIM_OBJ_SYNTAX, minim_builtin_delay);
-    minim_load_builtin(env, "force", MINIM_OBJ_FUNC, minim_builtin_force);
-    minim_load_builtin(env, "promise?", MINIM_OBJ_FUNC, minim_builtin_promisep);
-
-    // Port
-    minim_load_builtin(env, "current-input-port", MINIM_OBJ_FUNC, minim_builtin_current_input_port);
-    minim_load_builtin(env, "current-output-port", MINIM_OBJ_FUNC, minim_builtin_current_output_port);
-    minim_load_builtin(env, "call-with-input-file", MINIM_OBJ_FUNC, minim_builtin_call_with_input_file);
-    minim_load_builtin(env, "call-with-output-file", MINIM_OBJ_FUNC, minim_builtin_call_with_output_file);
-    minim_load_builtin(env, "with-input-from-file", MINIM_OBJ_FUNC, minim_builtin_with_input_from_file);
-    minim_load_builtin(env, "with-output-from-file", MINIM_OBJ_FUNC, minim_builtin_with_output_from_file);
-    minim_load_builtin(env, "port?", MINIM_OBJ_FUNC, minim_builtin_portp);
-    minim_load_builtin(env, "input-port?", MINIM_OBJ_FUNC, minim_builtin_input_portp);
-    minim_load_builtin(env, "output-port?", MINIM_OBJ_FUNC, minim_builtin_output_portp);
-    minim_load_builtin(env, "open-input-file", MINIM_OBJ_FUNC, minim_builtin_open_input_file);
-    minim_load_builtin(env, "open-output-file", MINIM_OBJ_FUNC, minim_builtin_open_output_file);
-    minim_load_builtin(env, "close-input-port", MINIM_OBJ_FUNC, minim_builtin_close_input_port);
-    minim_load_builtin(env, "close-output-port", MINIM_OBJ_FUNC, minim_builtin_close_output_port);
-
-    minim_load_builtin(env, "read", MINIM_OBJ_FUNC, minim_builtin_read);
-    minim_load_builtin(env, "read-char", MINIM_OBJ_FUNC, minim_builtin_read_char);
-    minim_load_builtin(env, "peek-char", MINIM_OBJ_FUNC, minim_builtin_peek_char);
-    minim_load_builtin(env, "char-ready?", MINIM_OBJ_FUNC, minim_builtin_char_readyp);
-
-    minim_load_builtin(env, "write", MINIM_OBJ_FUNC, minim_builtin_write);
-    minim_load_builtin(env, "display", MINIM_OBJ_FUNC, minim_builtin_display);
-    minim_load_builtin(env, "newline", MINIM_OBJ_FUNC, minim_builtin_newline);
-    minim_load_builtin(env, "write-char", MINIM_OBJ_FUNC, minim_builtin_write_char);
-
-    env_intern_sym(env, "eof", minim_eof);
-    minim_load_builtin(env, "eof?", MINIM_OBJ_FUNC, minim_builtin_eofp);
-}
-
 static void init_builtin(const char *name, MinimObjectType type, ...)
 {
     MinimObject *obj;
