+++ src/test/test-common.h
@@ -1,0 +1,10 @@
+#ifndef _MINIM_TEST_COMMON_H_
+#define _MINIM_TEST_COMMON_H_
+
+#include "../minim.h"
+
+void setup_test_env();
+bool run_test(char *input, char *expected);
+bool evaluate(char *input);
+
+#endif
