+++ tests/ref.c
@@ -2,31 +2,14 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
-#include "../src/minim.h"
-
-bool run_test(char *input, char *expected)
-{
-    char *str;
-    bool s;
-
-    str = eval_string(input, INT_MAX);
-    s = (strcmp(str, expected) == 0);
-    if (!s) printf("FAILED! input: %s, expected: %s, got: %s\n", input, expected, str);
-
-    return s;
-}
-
-bool evaluate(char *input)
-{
-    eval_string(input, INT_MAX);
-    return true;
-}
+#include "../src/test/test-common.h"
 
 int main()
 {
     bool status = true;
 
     GC_init(&status);
+    setup_test_env();
 
     {
         const int COUNT = 2;
